feat: add fields prevMonthDiff and prevMonthDiffPercentage to FipePrice model
add difference calculation to VehicleDetailsService#getVehicleById
